* releases: (133 commits)
  Changed solver from 'dense_cholesky' to 'cholesky' to eliminate deprecation warning.
  Added utility to skip tests if running on Travis
  MAINT: more explicit glob pattern in doc generation
  MAINT ensure that examples figures are displayed in the correct order
  #3356 - Added an exception raising when np.nan is passed into a HashingVectorizer.
  Corrected two typos in docstring.
  DOC updated installation documentation
  DOC: typo, envelop → envelope
  FIX #3372: unstable input check test for RANSACRegressor
  FIX multilabel deprecation warning in RidgeClassifierCV
  FIX assert_array_almost_equal for windows tests
  MAINT skip joblib multiprocessing tests on travis
  TST fix precision failure on windows
  DOC run optipng before uploading website
  Release 0.15.0b2
  MAINT remove residual sparsefuncs*.so when compiling
  DOC SVMs take string class labels as well as integers
  MAINT bump joblib to 0.8.2
  Convert windows_testing_downloader.ps1 from UTF-16le to UTF-8
  Metadata information for unpickling models in future versions
  ...

Conflicts:
	sklearn/externals/joblib/__init__.py
	sklearn/externals/joblib/parallel.py
	sklearn/externals/joblib/pool.py
	sklearn/externals/joblib/test/test_hashing.py
	sklearn/externals/joblib/test/test_logger.py
	sklearn/externals/joblib/test/test_pool.py

+++ sklearn/metrics/cluster/tests/test_supervised.py
@@ -189,7 +189,7 @@ def test_exactly_zero_info_score():
 
 def test_v_measure_and_mutual_information(seed=36):
     """Check relation between v_measure, entropy and mutual information"""
-    for i in np.logspace(1, 4, 4):
+    for i in np.logspace(1, 4, 4).astype(np.int):
         random_state = np.random.RandomState(seed)
         labels_a, labels_b = random_state.random_integers(0, 10, i),\
             random_state.random_integers(0, 10, i)
